Use <array> instead of <map> for powder storage
Random access of arrays is O(1) as opposed to O(logn) and should make
interactions more efficient since powders want to know what's around
them. However, this implementation is currently much slower because a
lot of copying of mostly empty arrays has to be done at the end of each
frame.