Improve powder "changed this frame" logic
Logic previously allowed a powder to have been changed during the
current frame or not. This meant that between frames, powders from last
frame had to be removed so as not to just appear later as powders that
had already been changed that frame.

Logic for invalidation is now based on current frame and the last frame
the powder was changed. No more additional invalidation between frames
removes the slowdown from having to traverse the powder data structure
again.